+++ lib/facebook/messenger.rb
@@ -1,9 +1,10 @@
 require 'facebook/messenger/version'
 require 'facebook/messenger/error'
 require 'facebook/messenger/subscriptions'
-require 'facebook/messenger/thread'
+require 'facebook/messenger/profile'
 require 'facebook/messenger/bot'
 require 'facebook/messenger/server'
+require 'facebook/messenger/server_no_error'
 require 'facebook/messenger/configuration'
 require 'facebook/messenger/incoming'