+++ lib/facebook/messenger/incoming/common.rb
@@ -21,7 +21,7 @@ module Facebook
           Time.at(@messaging['timestamp'] / 1000)
         end
 
-        def type
+        def typing_on
           payload = {
             recipient: sender,
             sender_action: 'typing_on'
@@ -30,10 +30,29 @@ module Facebook
           Facebook::Messenger::Bot.deliver(payload, access_token: access_token)
         end
 
+        def typing_off
+          payload = {
+            recipient: sender,
+            sender_action: 'typing_off'
+          }
+
+          Facebook::Messenger::Bot.deliver(payload, access_token: access_token)
+        end
+
+        def mark_seen
+          payload = {
+            recipient: sender,
+            sender_action: 'mark_seen'
+          }
+
+          Facebook::Messenger::Bot.deliver(payload, access_token: access_token)
+        end
+
         def reply(message)
           payload = {
             recipient: sender,
-            message: message
+            message: message,
+            message_type: Facebook::Messenger::Bot::MessageType::RESPONSE
           }
 
           Facebook::Messenger::Bot.deliver(payload, access_token: access_token)