+++ lib/facebook/messenger/incoming/postback.rb
@@ -8,6 +8,13 @@ module Facebook
         def payload
           @messaging['postback']['payload']
         end
+
+        def referral
+          return if @messaging['postback']['referral'].nil?
+          @referral ||= Referral::Referral.new(
+            @messaging['postback']['referral']
+          )
+        end
       end
     end
   end